Edit 12/27
Inserted a table for the meats without any visible borders. Will add another file with borders for the table.  (As of 12/27, this project is still in progress). For page two, I changed the color of the text to red to match the first page and I took out the underline of hyperlink--on both page--to make it more appealing to the user.